refactor(gui): Split gui.py into package with extracted infrastructure
Convert src/scanning_tool/gui.py into src/scanning_tool/gui/ package.
Extract infrastructure modules from the 807-line megafile:

- theme.py: apply_glass_theme, style_spinbox, GlassPalette type
- widgets.py: create_glass_scale, ScrollableFrame helper
- status.py: StatusBar class owning both status StringVars and
  anchor-hold/dedup logic (replaces scattered closures and dicts)
- alignment.py: AlignmentPoller encapsulates the after() polling loop
- lifecycle.py: register_close_handler for window teardown
- app.py: launch_gui entry point, now wired through the new helpers

Behavior unchanged; section-building code still lives inline in app.py
and will be extracted in follow-up milestones.
