+++ src/scanning_tool/gui/__init__.py
@@ -1,0 +1,5 @@
+"""Tkinter GUI package for the scanning tool."""
+
+from scanning_tool.gui.app import launch_gui
+
+__all__ = ["launch_gui"]
